refactor(core): extract trajectory types to separate module
- Create trajectory-types.ts with all type definitions (242 lines)
- trajectory.ts reduced from 1,762 to 1,578 lines
- Barrel export maintains full backward compatibility
- Zero external API changes

Part of v1.3 God file assessment.
